+++ translations/kartend_en.ts
@@ -5926,7 +5926,7 @@ Click Next to continue, or Cancel to skip — you can re-run this wizard any tim
 <context>
     <name>LauncherImportController</name>
     <message numerus="yes">
-        <location filename="../src/core/launcherimportcontroller.cpp" line="+166"/>
+        <location filename="../src/core/launcherimportcontroller.cpp" line="+189"/>
         <source>%1: re-synced — %2 added or updated, %3 removed, %n game(s) total.</source>
         <translation type="unfinished">
             <numerusform></numerusform>
@@ -5935,7 +5935,7 @@ Click Next to continue, or Cancel to skip — you can re-run this wizard any tim
     </message>
     <message numerus="yes">
         <location line="+6"/>
-        <location line="+48"/>
+        <location line="+49"/>
         <source>%1: Steam metadata filled for %n item(s).</source>
         <translation type="unfinished">
             <numerusform></numerusform>
@@ -5943,7 +5943,7 @@ Click Next to continue, or Cancel to skip — you can re-run this wizard any tim
         </translation>
     </message>
     <message numerus="yes">
-        <location line="-3"/>
+        <location line="-2"/>
         <source>%1: %n game(s) imported.</source>
         <translation type="unfinished">
             <numerusform></numerusform>
@@ -5951,7 +5951,7 @@ Click Next to continue, or Cancel to skip — you can re-run this wizard any tim
         </translation>
     </message>
     <message>
-        <location line="+13"/>
+        <location line="+14"/>
         <source>Import from Launcher</source>
         <translation type="unfinished"></translation>
     </message>